requirements.txt
absl-py==1.4.0
array-record==0.5.0
asttokens==2.4.1
astunparse==1.6.3
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
chex==0.1.85
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
decorator==5.1.1
dm-tree==0.1.8
etils==1.6.0
exceptiongroup==1.2.0
executing==2.0.1
filelock==3.13.1
flatbuffers==23.5.26
flax==0.7.5
fonttools==4.47.0
fsspec==2023.12.2
gast==0.5.4
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
googleapis-common-protos==1.62.0
grpcio==1.60.0
gviz-api==1.10.0
h5py==3.10.0
idna==3.6
imageio==2.33.1
importlib-resources==6.1.1
jax==0.4.23
jaxlib==0.4.23
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
keras==2.15.0
keras-tuner==1.4.6
kiwisolver==1.4.5
kt-legacy==1.0.5
libclang==16.0.6
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
memory-profiler==0.61.0
ml-dtypes==0.2.0
mpmath==1.3.0
msgpack==1.0.7
namex==0.0.7
nest-asyncio==1.5.8
networkx==3.2.1
numpy==1.26.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvcc-cu12==12.2.140
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.2.140
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
opencv-python==4.9.0.80
opt-einsum==3.3.0
optax==0.1.7
orbax-checkpoint==0.4.4
packaging==23.2
pandas==2.1.4
parso==0.8.3
pexpect==4.9.0
Pillow==10.1.0
promise==2.3
prompt-toolkit==3.0.43
protobuf==3.20.3
psutil==5.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
Pygments==2.17.2
pyparsing==3.1.1
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rsa==4.9
scikeras==0.12.0
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.1
six==1.16.0
stack-data==0.6.3
sympy==1.12
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorboard-plugin-profile==2.15.0
tensorflow==2.15.0.post1
tensorflow-addons==0.23.0
tensorflow-datasets==4.9.4
tensorflow-decision-forests==1.8.1
tensorflow-estimator==2.15.0
tensorflow-examples @ git+https://github.com/tensorflow/examples.git@fff4bcda7201645a1efaea4534403daf5fc03d42
tensorflow-hub==0.15.0
tensorflow-io-gcs-filesystem==0.35.0
tensorflow-metadata==1.14.0
tensorflow-model-optimization==0.7.5
tensorflowjs==4.15.0
tensorstore==0.1.45
termcolor==2.4.0
tf-keras==2.15.0
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
toml==0.10.2
toolz==0.12.0
torch==2.1.2
torchvision==0.16.2
tqdm==4.66.1
traitlets==5.14.0
triton==2.1.0
typeguard==2.13.3
typing_extensions==4.9.0
tzdata==2023.3
ultralytics==8.0.235
urllib3==2.1.0
wcwidth==0.2.12
Werkzeug==3.0.1
wrapt==1.14.1
wurlitzer==3.0.3
zipp==3.17.0
